MCU_43 - Moved icon creation into separate class for reuse
