+++ MoreCyclopsUpgrades/AuxConsole/AuxCyUpgradeConsoleMono.cs
@@ -1,9 +1,7 @@
 ﻿namespace MoreCyclopsUpgrades.AuxConsole
 {
-    using Common;
     using MoreCyclopsUpgrades.API.Buildables;
     using UnityEngine;
-    using UnityEngine.UI;
 
     internal class AuxCyUpgradeConsoleMono : AuxiliaryUpgradeConsole
     {
@@ -37,43 +35,12 @@
 
             var rotation = Quaternion.Euler(60f, 180, 0);
 
-            ModuleDisplay1 = CreateModuleDisplay(new Vector3(rightColX, botRowY, botRowZ), rotation);
-            ModuleDisplay2 = CreateModuleDisplay(new Vector3(middColX, botRowY, botRowZ), rotation);
-            ModuleDisplay3 = CreateModuleDisplay(new Vector3(leftColX, botRowY, botRowZ), rotation);
-            ModuleDisplay4 = CreateModuleDisplay(new Vector3(rightColX, topRowY, topRowZ), rotation);
-            ModuleDisplay5 = CreateModuleDisplay(new Vector3(middColX, topRowY, topRowZ), rotation);
-            ModuleDisplay6 = CreateModuleDisplay(new Vector3(leftColX, topRowY, topRowZ), rotation);
-        }
-
-        private GameObject CreateModuleDisplay(Vector3 position, Quaternion rotation)
-        {
-            const float scale = 0.215f;
-
-            Canvas canvas = new GameObject("Canvas", typeof(RectTransform)).AddComponent<Canvas>();
-            Transform t = canvas.transform;
-            t.SetParent(this.transform, false);
-            canvas.sortingLayerID = 1;
-
-            uGUI_GraphicRaycaster raycaster = canvas.gameObject.AddComponent<uGUI_GraphicRaycaster>();
-
-            var rt = t as RectTransform;
-            RectTransformExtensions.SetParams(rt, new Vector2(0.5f, 0.5f), new Vector2(0.5f, 0.5f));
-            RectTransformExtensions.SetSize(rt, 0.5f, 0.5f);
-
-            t.localPosition = position;
-            t.localRotation = rotation;
-            t.localScale = new Vector3(scale, scale, scale);
-
-            canvas.scaleFactor = 0.01f;
-            canvas.renderMode = RenderMode.WorldSpace;
-            canvas.referencePixelsPerUnit = 100;
-
-            CanvasScaler scaler = canvas.gameObject.AddComponent<CanvasScaler>();
-            scaler.dynamicPixelsPerUnit = 20;
-
-            uGUI_Icon icon = canvas.gameObject.AddComponent<uGUI_Icon>();
-            icon.enabled = false;
-            return canvas.gameObject;
+            ModuleDisplay1 = IconCreator.CreateModuleDisplay(this, new Vector3(rightColX, botRowY, botRowZ), rotation);
+            ModuleDisplay2 = IconCreator.CreateModuleDisplay(this, new Vector3(middColX, botRowY, botRowZ), rotation);
+            ModuleDisplay3 = IconCreator.CreateModuleDisplay(this, new Vector3(leftColX, botRowY, botRowZ), rotation);
+            ModuleDisplay4 = IconCreator.CreateModuleDisplay(this, new Vector3(rightColX, topRowY, topRowZ), rotation);
+            ModuleDisplay5 = IconCreator.CreateModuleDisplay(this, new Vector3(middColX, topRowY, topRowZ), rotation);
+            ModuleDisplay6 = IconCreator.CreateModuleDisplay(this, new Vector3(leftColX, topRowY, topRowZ), rotation);
         }
 
         private void SetModuleVisibility(string slot, bool visible)
