+++ MoreCyclopsUpgrades/IconCreator.cs
@@ -1,0 +1,75 @@
+﻿namespace MoreCyclopsUpgrades
+{
+    using System;
+    using System.Collections.Generic;
+    using MoreCyclopsUpgrades.AuxConsole;
+    using MoreCyclopsUpgrades.Managers;
+    using UnityEngine;
+    using UnityEngine.UI;
+
+    internal class IconCreator
+    {
+        internal static GameObject CreateModuleDisplay(AuxCyUpgradeConsoleMono upgradeConsole, Vector3 position, Quaternion rotation)
+        {
+            const float scale = 0.215f;
+
+            Canvas canvas = new GameObject("Canvas", typeof(RectTransform)).AddComponent<Canvas>();
+            Transform t = canvas.transform;
+            t.SetParent(upgradeConsole.transform, false);
+            canvas.sortingLayerID = 1;
+
+            uGUI_GraphicRaycaster raycaster = canvas.gameObject.AddComponent<uGUI_GraphicRaycaster>();
+
+            var rt = t as RectTransform;
+            RectTransformExtensions.SetParams(rt, new Vector2(0.5f, 0.5f), new Vector2(0.5f, 0.5f));
+            RectTransformExtensions.SetSize(rt, 0.5f, 0.5f);
+
+            t.localPosition = position;
+            t.localRotation = rotation;
+            t.localScale = new Vector3(scale, scale, scale);
+
+            canvas.scaleFactor = 0.01f;
+            canvas.renderMode = RenderMode.WorldSpace;
+            canvas.referencePixelsPerUnit = 100;
+
+            CanvasScaler scaler = canvas.gameObject.AddComponent<CanvasScaler>();
+            scaler.dynamicPixelsPerUnit = 20;
+
+            uGUI_Icon icon = canvas.gameObject.AddComponent<uGUI_Icon>();
+            icon.enabled = false;
+            return canvas.gameObject;
+        }
+
+        internal static PowerIndicatorIcon CreatePowerIndicatorIcon(Canvas canvas, float xoffset, float yoffset, float zoffset, float scale)
+        {
+            var iconGo = new GameObject("IconGo", typeof(RectTransform));
+            iconGo.transform.SetParent(canvas.transform, false);
+            iconGo.transform.localPosition = new Vector3(xoffset, yoffset, zoffset);
+            iconGo.transform.localScale = new Vector3(scale, scale, scale);
+
+            uGUI_Icon icon = iconGo.AddComponent<uGUI_Icon>();
+
+            var arial = (Font)Resources.GetBuiltinResource(typeof(Font), "Arial.ttf");
+            var textGO = new GameObject("TextGo");
+
+            textGO.transform.SetParent(iconGo.transform, false);
+
+            Text text = textGO.AddComponent<Text>();
+            text.font = arial;
+            text.material = arial.material;
+            text.text = "??";
+            text.fontSize = 22;
+            text.alignment = TextAnchor.LowerCenter;
+            text.color = Color.white;
+
+            Outline outline = textGO.AddComponent<Outline>();
+            outline.effectColor = Color.black;
+
+            RectTransform rectTransform = text.GetComponent<RectTransform>();
+            rectTransform.localScale = Vector3.one;
+            rectTransform.anchoredPosition3D = Vector3.zero;
+            rectTransform.anchoredPosition += new Vector2(0f, -15f);
+            return new PowerIndicatorIcon(icon, text);
+        }
+    }
+}
